# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Pawn Broker API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /loans:
    get:
      operationId: listLoans
      tags: [loans]
      summary: List loans
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: loanType
          in: query
          schema:
            type: string
        - name: customerId
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Loan list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoanListResponse"
    post:
      operationId: createLoan
      tags: [loans]
      summary: Create loan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoanInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
  /loans/{id}:
    get:
      operationId: getLoan
      tags: [loans]
      summary: Get loan details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Loan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoanDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updateLoan
      tags: [loans]
      summary: Update loan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoanUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
  /loans/{id}/payment:
    post:
      operationId: recordPayment
      tags: [loans]
      summary: Record a payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Payment recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /loans/{id}/close:
    post:
      operationId: closeLoan
      tags: [loans]
      summary: Close a loan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoanCloseInput"
      responses:
        "200":
          description: Loan closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
  /loans/{id}/renew:
    post:
      operationId: renewLoan
      tags: [loans]
      summary: Renew a loan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoanRenewalInput"
      responses:
        "200":
          description: Loan renewed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
...
components:
  schemas:
...
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    JewelleryItem:
      type: object
      required: [id, loanId, jewelleryType, grossWeight, netWeight, purity, estimatedValue]
      properties:
        id:
          type: integer
        loanId:
          type: integer
        jewelleryType:
          type: string
        category:
          type: ["string", "null"]
        grossWeight:
          type: number
        stoneWeight:
          type: number
        netWeight:
          type: number
        purity:
          type: string
        estimatedValue:
          type: number
        marketValue:
          type: ["number", "null"]
    JewelleryItemInput:
      type: object
      required: [jewelleryType, grossWeight, purity, estimatedValue]
      properties:
        jewelleryType:
          type: string
        category:
          type: string
        grossWeight:
          type: number
        stoneWeight:
          type: number
        netWeight:
          type: number
        purity:
          type: string
        estimatedValue:
          type: number
        marketValue:
          type: number
    Loan:
      type: object
      required: [id, loanNumber, customerId, loanType, principalAmount, interestRate, ratePeriod, interestType, status, createdAt]
      properties:
        id:
          type: integer
        loanNumber:
          type: string
        customerId:
          type: integer
        customerName:
          type: ["string", "null"]
        loanType:
          type: string
          enum: [gold, silver]
        principalAmount:
          type: number
        interestRate:
          type: number
        ratePeriod:
          type: string
          enum: [day, month, year]
        interestType:
          type: string
          enum: [simple, compound]
        duration:
          type: integer
        durationUnit:
          type: string
          enum: [days, months, years]
        startDate:
          type: string
        dueDate:
          type: ["string", "null"]
        expiryDate:
          type: ["string", "null"]
        totalInterest:
          type: number
        totalPayable:
          type: number
        amountPaid:
          type: number
        outstandingBalance:
          type: number
        penaltyRate:
          type: ["number", "null"]
        status:
          type: string
          enum: [active, closed, overdue, auction, partially_paid]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    LoanDetail:
      type: object
      required: [id, loanNumber, customerId, loanType, principalAmount, interestRate, ratePeriod, interestType, status, createdAt, jewelleryItems, payments]
      properties:
        id:
          type: integer
        loanNumber:
          type: string
        customerId:
          type: integer
        customerName:
          type: ["string", "null"]
        customerPhone:
          type: ["string", "null"]
        loanType:
          type: string
          enum: [gold, silver]
        principalAmount:
          type: number
        interestRate:
          type: number
        ratePeriod:
          type: string
          enum: [day, month, year]
        interestType:
          type: string
          enum: [simple, compound]
        duration:
          type: integer
        durationUnit:
          type: string
          enum: [days, months, years]
        startDate:
          type: string
        dueDate:
          type: ["string", "null"]
        expiryDate:
          type: ["string", "null"]
        totalInterest:
          type: number
        totalPayable:
          type: number
        amountPaid:
          type: number
        outstandingBalance:
          type: number
        penaltyRate:
          type: ["number", "null"]
        status:
          type: string
          enum: [active, closed, overdue, auction, partially_paid]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        jewelleryItems:
          type: array
          items:
            $ref: "#/components/schemas/JewelleryItem"
        payments:
          type: array
          items:
            $ref: "#/components/schemas/Payment"
    LoanInput:
      type: object
      required: [customerId, loanType, principalAmount, interestRate, ratePeriod, interestType, duration, durationUnit, jewelleryItems]
      properties:
        customerId:
          type: integer
        loanType:
          type: string
          enum: [gold, silver]
        principalAmount:
          type: number
        interestRate:
          type: number
        ratePeriod:
          type: string
          enum: [day, month, year]
        interestType:
          type: string
          enum: [simple, compound]
        duration:
          type: integer
        durationUnit:
          type: string
          enum: [days, months, years]
        penaltyRate:
          type: number
        notes:
          type: string
        jewelleryItems:
          type: array
          items:
            $ref: "#/components/schemas/JewelleryItemInput"
    LoanUpdate:
      type: object
      properties:
        status:
          type: string
        notes:
          type: string
        penaltyRate:
          type: number
    LoanCloseInput:
      type: object
      required: [paymentMode]
      properties:
        discount:
          type: number
        paymentMode:
          type: string
          enum: [cash, upi, bank_transfer, cheque]
        referenceNumber:
          type: string
        notes:
          type: string
    LoanRenewalInput:
      type: object
      required: [newInterestRate, newDuration, newDurationUnit]
      properties:
        newInterestRate:
          type: number
        newDuration:
          type: integer
        newDurationUnit:
          type: string
          enum: [days, months, years]
        capitalizeInterest:
          type: boolean
        remarks:
          type: string
    LoanListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Loan"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    Payment:
      type: object
      required: [id, receiptNumber, loanId, paymentDate, amount, paymentMode, createdAt]
      properties:
        id:
          type: integer
        receiptNumber:
          type: string
        loanId:
          type: integer
        paymentDate:
          type: string
        amount:
          type: number
        interestPaid:
          type: number
        principalPaid:
          type: number
        penaltyPaid:
          type: number
        remainingPrincipal:
          type: number
        remainingInterest:
          type: number
        paymentMode:
          type: string
          enum: [cash, upi, bank_transfer, cheque]
        referenceNumber:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    PaymentInput:
      type: object
      required: [amount, paymentMode, paymentDate]
      properties:
        paymentDate:
          type: string
        amount:
          type: number
        interestPaid:
          type: number
        principalPaid:
          type: number
        penaltyPaid:
          type: number
        paymentMode:
          type: string
          enum: [cash, upi, bank_transfer, cheque]
        referenceNumber:
          type: string
        notes:
          type: string
...